added response body logging
